Clean up project
Removed the untested IRC widget, will commit it when it has actually been tested,
the logger is a bit more fleshed out.  Wrote skeletons for sample plugins to handle
the three main types of plugins (communication, VCS integration, and CI/builders).
Scrapped the idea of projects (for now) and each instance will handle a single
project.